feat(fees-subscription): add fee for subscription flow; test; minor refactoring
